admin-ui/pyproject.toml
[project]
name = "powerblockade-admin"
version = "0.1.0"
description = "PowerBlockade Admin UI - DNS control plane"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.2",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==3.2.2",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "orjson>=3.10",
    "dnspython>=2.6",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.26",
    "ruff>=0.6",
    "pytest-playwright>=0.5",
    "pytest-asyncio>=0.21",
    "pyright>=1.1",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring database",
    "e2e: End-to-end tests requiring full stack",
    "playwright: Browser tests requiring Playwright",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
